services/multimodal_service.py
import os
import logging
from typing import Dict, Any, List, Optional
from .base_service import BaseService
import io
import base64

try:
    import pdfplumber
    import PyPDF2
    from PIL import Image
    import cv2
    import numpy as np
except ImportError as e:
    logging.warning(f"Some dependencies not available: {e}")

class MultimodalService(BaseService):
    """多模态服务 - 处理PDF和图片文件转换为文本"""
    
    def __init__(self):
        super().__init__()
        self.supported_image_types = ['png', 'jpg', 'jpeg']
        self.supported_document_types = ['pdf']
    
    def process_pdf(self, notice_id: str, pdf_data: bytes, extra_info: Dict[str, Any]) -> Dict[str, Any]:
        """
        处理PDF文件转换为文本
        
        Args:
            notice_id: 公告ID
            pdf_data: PDF文件二进制数据
            extra_info: 额外信息字典
            
        Returns:
            处理结果
        """
        try:
            # TODO: 实现PDF文本提取逻辑
            # 这里可以使用pdfplumber或PyPDF2来提取文本
            
            result = {
                'notice_id': notice_id,
                'service_type': 'multimodal_pdf',
                'file_type': 'pdf',
                'extracted_text': '待实现 - PDF文本提取',
                'page_count': '待实现',
                'file_size': len(pdf_data)
            }
            
            return self.format_response(True, data=result)
            
        except Exception as e:
            self.logger.error(f"Error processing PDF: {str(e)}")
            return self.format_response(False, error=str(e))
    
    def process_images(self, notice_id: str, image_files: List[bytes], extra_info: Dict[str, Any]) -> Dict[str, Any]:
        """
        处理图片文件转换为文本
        
        Args:
            notice_id: 公告ID
            image_files: 图片文件二进制数据列表
            extra_info: 额外信息字典
            
        Returns:
            处理结果
        """
        try:
            # TODO: 实现图片OCR文本提取逻辑
            # 这里可以使用OCR技术（如Tesseract、PaddleOCR等）来提取文本
            
            extracted_texts = []
            for i, img_data in enumerate(image_files):
                # 模拟OCR处理
                extracted_texts.append(f"待实现 - 图片{i+1}的OCR文本提取")
            
            result = {
                'notice_id': notice_id,
                'service_type': 'multimodal_images',
                'file_type': 'images',
                'image_count': len(image_files),
                'extracted_texts': extracted_texts,
                'combined_text': ' '.join(extracted_texts)
            }
            
            return self.format_response(True, data=result)
            
        except Exception as e:
            self.logger.error(f"Error processing images: {str(e)}")
            return self.format_response(False, error=str(e))
    
    def validate_pdf(self, pdf_data: bytes) -> bool:
        """
        验证PDF文件
        
        Args:
            pdf_data: PDF文件二进制数据
            
        Returns:
            验证是否通过
        """
        if not pdf_data or len(pdf_data) == 0:
            return False
        
        # 检查PDF文件头
        if not pdf_data.startswith(b'%PDF'):
            return False
        
        return True
    
    def validate_image(self, image_data: bytes) -> bool:
        """
        验证图片文件
        
        Args:
            image_data: 图片文件二进制数据
            
        Returns:
            验证是否通过
        """
        if not image_data or len(image_data) == 0:
            return False
        
        try:
            # 尝试打开图片验证格式
            image = Image.open(io.BytesIO(image_data))
            image.verify()
            return True
        except Exception:
            return False
    
    def get_file_extension(self, filename: str) -> str:
        """
        获取文件扩展名
        
        Args:
            filename: 文件名
            
        Returns:
            文件扩展名（小写）
        """
        return os.path.splitext(filename)[1].lower().lstrip('.')
    
    def process(self, notice_id: str, content: str, extra_info: Dict[str, Any]) -> Dict[str, Any]:
        """
        基础处理方法（继承自BaseService，但多模态服务主要使用专门的方法）
        
        Args:
            notice_id: 公告ID
            content: 文本内容（多模态服务中可能不使用）
            extra_info: 额外信息字典
            
        Returns:
            处理结果
        """
        return self.format_response(False, error="MultimodalService should use process_pdf or process_images methods") 