requirements.txt
tornado>=6.1
python-multipart>=0.0.5
Pillow>=9.0.0
PyPDF2>=3.0.0
pdfplumber>=0.7.0
opencv-python>=4.5.0
numpy>=1.21.0 